refactor: Move webview/ to webview/setting/ to make room for sync UI
