+++ webview/setting/locales/ja.ts
